feat(feature/#23): 대사박스 수정, 피드백
