+++ src/styles/globals.css
@@ -29,7 +29,3 @@
   font-weight: normal;
   font-style: normal;
 }
-
-.timeline-description {
-  white-space: pre-line;
-}
